updated parser logic; added "clear_caches" button to admin panel;

+++ apps/rcoi/models.py
@@ -1,11 +1,36 @@
 import logging
+import datetime
 from django.core.urlresolvers import reverse
 from django.db import connection, models
 from django_extensions.db.models import TimeStampedModel
 
-from apps.rcoi.xlsx_to_csv import get_files, save_to_csv
+from apps.rcoi import xlsx_to_csv
 
 logger = logging.getLogger(__name__)
+
+
+class DataSource(TimeStampedModel):
+    name = models.CharField('Название', max_length=50)
+    url = models.URLField('Ссылка на источник данных')
+
+    class Meta:
+        ordering = ['-modified']
+
+    def __str__(self):
+        return str(self.name)
+
+
+class DataFile(TimeStampedModel):
+    name = models.CharField('Имя файла', max_length=50)
+    url = models.URLField('Ссылка на файл', unique=True)
+    size = models.IntegerField('Content-Length', blank=True, null=True)
+    last_modified = models.DateTimeField('Last-Modified', blank=True, null=True)
+
+    class Meta:
+        ordering = ['-last_modified']
+
+    def __str__(self):
+        return str(self.name)
 
 
 class Date(TimeStampedModel):
@@ -133,6 +158,7 @@ class Exam(TimeStampedModel):
     place = models.ForeignKey(Place, related_name='exams', on_delete=models.CASCADE)
     employee = models.ForeignKey(Employee, related_name='exams', on_delete=models.CASCADE)
     position = models.ForeignKey(Position, related_name='exams', on_delete=models.CASCADE)
+    datafile = models.ForeignKey(DataFile, related_name='exams', on_delete=models.CASCADE)
 
     class Meta:
         unique_together = (('date', 'level', 'place', 'employee'),)
@@ -148,21 +174,18 @@ class Exam(TimeStampedModel):
         return reverse('rcoi:exam_update', args=(self.id,))
 
 
-class DataSource(TimeStampedModel):
-    name = models.CharField('Название', max_length=50)
-    url = models.URLField('Ссылка на источник данных')
-
-
 class RcoiUpdater:
     def __init__(self):
         self.data = self.__prepare_data()
 
     def run(self):
-        self.__update_simple_tables()
-        self.__update_territory()
-        self.__update_employee()
-        self.__update_place()
-        self.__update_exam()
+        if self.data:
+            self.__update_simple_tables()
+            self.__update_territory()
+            self.__update_employee()
+            self.__update_place()
+            self.__update_exam()
+            self.__cleanup()
 
     def __prepare_data(self):
         import csv
@@ -171,26 +194,63 @@ class RcoiUpdater:
         from collections import defaultdict
 
         urls = DataSource.objects.all()
-        data = defaultdict(list)
         tmp_path = 'tmp'
-        csv_file = os.path.join(tmp_path, tmp_path + '.csv')
 
         if not os.path.exists(tmp_path):
             os.makedirs(tmp_path)
-        if not os.path.isfile(csv_file):
-            open(csv_file, 'w+').close()
 
-        [get_files(url.url, tmp_path) for url in urls]
-        save_to_csv(csv_file)
+        files_info = [xlsx_to_csv.get_files_info(url.url) for url in urls]
+        files_info = [url for url_list in files_info for url in url_list]
 
-        with open(csv_file, encoding='utf-8') as f:
-            reader = csv.DictReader(f, delimiter=';')
+        updated_files = []
+        for file in files_info:
+            name = file['name']
+            try:
+                f = DataFile.objects.get(name=name)
+                if f.last_modified != file['last_modified']:
+                    for k, v in file.items():
+                        setattr(f, k, v)
+                    f.save()
+                    updated_files.append(f)
+                    logger.debug('%s: dates differ: download', name)
+                else:
+                    logger.debug('%s: dates are the same: skip', name)
+            except DataFile.DoesNotExist:
+                logger.debug('%s: datafile not found: add and download', name)
+                f = DataFile.objects.create(**file)
+                updated_files.append(f)
+
+        if updated_files:
+            data = defaultdict(list)
+
+            [xlsx_to_csv.download_file(file.url, tmp_path) for file in updated_files]
+            csv_stream = xlsx_to_csv.save_to_stream(tmp_path)
+
+            reader = csv.DictReader(csv_stream, delimiter='\t')
             for row in reader:
                 for (k, v) in row.items():
                     data[k].append(v)
-        logger.debug('cleanup downloaded files...')
-        shutil.rmtree(tmp_path)
-        return data
+            logger.debug('cleanup downloaded files')
+            shutil.rmtree(tmp_path)
+            return data
+        return None
+
+    def __cleanup(self):
+        from django.conf import settings
+
+        logger.debug('cleanup unneeded exam rows')
+        now = datetime.datetime.now() - datetime.timedelta(minutes=20)
+        updated_files = DataFile.objects.filter(modified__gt=now)
+        for file in updated_files:
+            filtered = Exam.objects.filter(modified__lt=now, datafile_id=file.id)
+            filtered.delete()
+            logger.debug('%s: deleted %s rows', file.name, filtered.count())
+
+        if not settings.DEBUG:
+            from cacheops import invalidate_all
+            from django.core.cache import cache
+            invalidate_all()
+            cache.clear()
 
     def __sql_insert_or_update(self, table, columns, data, uniq):
         table_name = 'rcoi_' + table
@@ -214,7 +274,7 @@ class RcoiUpdater:
             if key == 'date' or key == 'level':
                 col = key
             columns = (col, 'created', 'modified')
-            logger.debug('processing %s...', table)
+            logger.debug('processing model: %s', table)
             self.__sql_insert_or_update(table, columns, stream, col)
 
     def __update_territory(self):
@@ -226,7 +286,7 @@ class RcoiUpdater:
         stream = timestamp_list(values)
         table = 'territory'
         columns = ('code', 'name', 'created', 'modified')
-        logger.debug('processing %s...', table)
+        logger.debug('processing model: %s', table)
         self.__sql_insert_or_update(table, columns, stream, columns[0])
 
     def __update_employee(self):
@@ -239,7 +299,7 @@ class RcoiUpdater:
         stream = timestamp_list(values_with_id)
         table = 'employee'
         columns = ('name', 'org_id', 'created', 'modified')
-        logger.debug('processing %s...', table)
+        logger.debug('processing model: %s', table)
         self.__sql_insert_or_update(table, columns, stream, columns[:2])
 
     def __update_place(self):
@@ -257,12 +317,10 @@ class RcoiUpdater:
         stream = timestamp_list(values_with_id)
         table = 'place'
         columns = ('code', 'name', 'addr', 'ate_id', 'created', 'modified')
-        logger.debug('processing %s...', table)
+        logger.debug('processing model: %s', table)
         self.__sql_insert_or_update(table, columns, stream, columns[:4])
 
     def __update_exam(self):
-        from datetime import timedelta
-
         date = Date.objects.all()
         date_db = {str(d.date): d.id for d in date}
         date_id = self.data['date'][:]
@@ -291,22 +349,22 @@ class RcoiUpdater:
                                self.data['organisation'], ))
         replace_items(employee_id, employee_db)
 
-        exams = list(zip(date_id, level_id, place_id, employee_id, position_id))
+        datafile = DataFile.objects.all()
+        datafile_db = {df.name: df.id for df in datafile}
+        datafile_id = self.data['datafile'][:]
+        replace_items(datafile_id, datafile_db)
+
+        exams = list(zip(date_id, level_id, place_id, employee_id, position_id, datafile_id))
 
         table = 'exam'
         columns = ('date_id', 'level_id', 'place_id',
-                   'employee_id', 'position_id',
+                   'employee_id', 'position_id', 'datafile_id',
                    'created', 'modified')
-        logger.debug('processing %s...', table)
+        logger.debug('processing model: %s', table)
 
         for chunk in split_list(exams, 20):
             stream = timestamp_list(chunk)
             self.__sql_insert_or_update(table, columns, stream, columns[:4])
-
-        # Delete unmodified rows ( == was deleted from files)
-        for model in (Exam, Employee, Place):
-            now = model.objects.latest('modified').modified - timedelta(minutes=5)
-            model.objects.filter(modified__lt=now).delete()
 
 
 def replace_items(s_list, s_dict):
@@ -315,7 +373,6 @@ def replace_items(s_list, s_dict):
 
 
 def timestamp_list(data):
-    import datetime
     datetime_now = datetime.datetime.now()
     created = [datetime_now, datetime_now]
     data_list = []
